requirements.txt
Django>=4.0,<5.0
mysqlclient>=2.1.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
python-dotenv>=1.0.0
Pillow>=10.0.0
